Create a ReaderWriterLockAsync class
The original intention was to allow the .NET 6 implementation of
DiskWriter to write multiple blocks to the same underlying file handle
at the same time, as that is one of the supported capabilities of the
RandomAccess class.

This could have been supported by using a ReaderWriterLock (which was
compatible with async/await) so that closing the handle required the
writer lock, but reading/writing blocks of data would use the Reader
lock.

However, this greatly complicated the implementation based on FileStream
which was used for .NET 5.0 and older.

To keep it simple, the approach was altered to just create multiple
handles instead of using a readerwriterlock.

However, I like the implementation so i may use it for something...
eventually.